refactor: ExpoData프로토콜 구현 및 Json관련 메서드 분리
- Exposition, ExpoItemElement가 ExpoData 프로토콜을 준수하도록 변경
- loadJsonData, loadFileLocation 메서드를 재사용하기 위해 리팩토링
- ExpoHomeViewController에서 Json 데이터를 디코딩 하는 메서드를 분리
