+++ Expo1900/Expo1900/Controller/ExpoHomeViewController.swift
@@ -21,7 +21,7 @@ class ExpoHomeViewController: UIViewController {
     
     override func viewDidLoad() {
         super.viewDidLoad()
-        loadJsonData("exposition", "json")
+        setExpositionData()
         updateUI()
     }
     
@@ -42,22 +42,8 @@ class ExpoHomeViewController: UIViewController {
         self.navigationController?.pushViewController(menuVC, animated: true)
     }
     
-    private func loadJsonData(_ fileName: String, _ extensionName: String) {
-        let jsonDecoder = JSONDecoder()
-        let fileUrl = loadFileLocation(fileName, extensionName)
-        do {
-            let data = try Data(contentsOf: fileUrl)
-            let decodedData = try jsonDecoder.decode(Exposition.self, from: data)
-            exposition = decodedData
-        } catch {
-            fatalError("JSON DATA LOAD ERROR: \(fileName),\(extensionName)")
-        }
-    }
-    
-    private func loadFileLocation(_ fileName: String, _ extensionName: String) -> URL {
-        guard let fileLocation = Bundle.main.url(forResource: fileName, withExtension: extensionName) else {
-            return URL(fileURLWithPath: "")
-        }
-        return fileLocation
+    private func setExpositionData() {
+        let exposition = loadJsonData(type: Exposition.self, "exposition", "json") 
+        self.exposition = exposition
     }
 }
