+++ Expo1900/Expo1900/Protocol/ExpoData.swift
@@ -1,0 +1,10 @@
+//
+//  ExpoData.swift
+//  Expo1900
+//
+//  Created by 유한석 on 2022/06/17.
+//
+
+protocol ExpoData {
+    
+}
